+++ __init__.py
@@ -19,6 +19,8 @@ import traceback
 NODE_MODULES = [
     "node_image_crop_rotate_pad",
     "node_lama_inpaint",
+    "node_load_video",
+    "node_refine_mask",
     "node_select_frame",
     "node_video_crop_rotate_pad",
 ]